+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Fixed
+
+- Restored deprecated `setfile` method with deprecation warning to change to `writefile`
+
 ## [2.4.4] - 2019-02-23
 
 ### Fixed